Square 분리
